eal: force compilation of RTE_ASSERT expression
Even if RTE_ENABLE_ASSERT is not enabled, the expression used should
still be checked for compiler warnings. Use sizeof()
and ternary operator in same manner as the assert() macro to
cause the expression to be evaluated but not generate code.

This was motivated by detection of problems in a driver submission
that was not detected until compiled with non default flags.

Signed-off-by: Stephen Hemminger <stephen@networkplumber.org>
Signed-off-by: 0-day Robot <robot@bytheb.org>