net/mlx5: fix the GRE mask set on root table
When setting the value, the GRE mask cannot reuse the one from the
value, or else the value cannot be really ANDed with the proper input
mask from the user. If the value contains more valid bits than the
mask, the rule insertion on the root table will get a failure due to
the extra bits in the value field.

Using the input mask or the default mask will help to clear the extra
bits and solve the issue.

Fixes: 25ab2cbba31d ("net/mlx5: fix GRE item translation for root table")

cc: stable@dpdk.org

Signed-off-by: Bing Zhao <bingz@nvidia.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>