fbarray: flatten the loop in find next n
Currently, find_next_n() is implemented as a nested loop due to lookahead
functionality. This is not very efficient because when doing lookahead,
we essentially scan some of the indices twice, and in general the
lookahead functionality has been a source of bugs because it is
overcomplicated. The bit ignore feature on lookahead is also unnecessary
because we don't really gain anything by ignoring bits we have already
scanned, as they would not trigger any matches anyway.

This patch reworks find_next_n() to flatten the loop and remove the
lookahead and bit-ignore functionality, instead replacing it with state
machine-like behavior. This makes the code simpler to reason about.

Signed-off-by: Anatoly Burakov <anatoly.burakov@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>