fbarray: flatten the loop in find prev n
Currently, find_prev_n() is implemented as a nested loop due to
lookbehind functionality. This is not very efficient because when doing
lookbehind, we essentially scan some of the indices twice, and in general
the lookbehind functionality has been a source of bugs because it is
overcomplicated. The bit ignore feature on lookbehind is also unnecessary
because we don't really win anything by ignoring bits we have already
scanned, as they would not trigger any matches anyway.

This patch reworks find_prev_n() to flatten the loop and remove the
lookbehind and bit-ignore functionality, instead replacing it with state
machine-like behavior. This makes the code simpler to reason about.

Signed-off-by: Anatoly Burakov <anatoly.burakov@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>