net/mlx5: fix LACP packets handling in isolated mode
In HWS(dv_flow_en=2) mode, in the routine that enables traffic,  isolated
mode check was before LACP default rules creation as opposed to
legacy mode handling.

As a result, the LACP default rules were not created in HWS mode
and LACP packets arrived to testpmd.
The fix was to create the LACP default rules in isolated mode.

Fixes: 87e4384d2662 ("net/mlx5: fix condition of LACP miss flow")
Cc: stable@dpdk.org

Signed-off-by: Maayan Kashani <mkashani@nvidia.com>
Acked-by: Dariusz Sosnowski <dsosnowski@nvidia.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>