net/ice: fix flow engines order
Reorder flow engine evaluation for RTE flow rule installation.
Currently, the driver evaluates rules with engines in the order they are
executed in the hardware. However, in this situation, some wildcarded flows
that are also supported by the ACL engine are installed as switch engine
rules, which scales poorly with each distinct wildcard mask configuration.

Fixes: fabc9e1322e2 ("net/ice: fix flows handling")
Cc: stable@dpdk.org

Signed-off-by: Vladimir Medvedkin <vladimir.medvedkin@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>