ci: fix ccache for Ubuntu 22.04
By default (and if no $HOME/.ccache directory exists), Ubuntu 22.04 ccache
stores its data in $HOME/.cache/ccache.
As a result, since the switch to Ubuntu 22.04, ccache has not been
functional.

Set the ccache directory to the location cached in GHA.

Note: ccache is working with the Fedora containers, as the presence of a
.ccache directory made ccache use it, still,
explicitly set CCACHE_DIR for consistency.

Fixes: 8400093195dd ("ci: bump tested distributions in GHA")
Cc: stable@dpdk.org

Signed-off-by: David Marchand <david.marchand@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>