net/mlx5: fix mark flow action validation in FDB mode
The MARK flow action is supported in FDB SWS mode, but it
was not supported in the early FDB HWS implementations.

The current MARK flow action validation reflects
the original HWS functionality.

Recent MLX5 FW supports the MARK flow action in FDB mode.

The patch clears the MARK action for both SWS FDB and HWS FDB.

Cc: stable@dpdk.org

Fixes: d6dc072aeb12 ("net/mlx5: validate flow actions in table creation")

Signed-off-by: Gregory Etelson <getelson@nvidia.com>
Acked-by: Dariusz Sosnowski <dsosnowski@nvidia.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>