net/mlx5: fix NAT64 register selection
PMD statically assumed that REG_C_6 is always available for use
with NAT64 HW flow action.
This led to PMD configuration errors on FW versions which do not expose
that specific register.

This patch fixes that by adding a check for REG_C_6 against FW
capabilities, when registers for NAT64 are selected.

Also, if not enough registers are available, PMD will not attempt to
create NAT64 actions.

Fixes: 7a26bfec06a4 ("net/mlx5: fetch available registers for NAT64")
Cc: stable@dpdk.org

Signed-off-by: Dariusz Sosnowski <dsosnowski@nvidia.com>
Acked-by: Bing Zhao <bingz@nvidia.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>