dts: add Scapy asynchronous sniffer
Currently packet capturing in Scapy is used by running Scapy's own
class AsyncSniffer in the same shell as packet sending. Because there is
some start up time involved in the AsyncSniffer, doing this every single
time a test case requires it can become slow and unreliable if sniffer
was not ready yet at the time of sending.

Add a multi-threaded implementation for sniffing and capturing packets
on-demand using Scapy. The sniffer lives in its own dedicated shell,
which is setup for sniffing at the start up of the test run. Packets are
only captured when requested by the tests through the use of Event
signals.

Signed-off-by: Luca Vizzarro <luca.vizzarro@arm.com>
Reviewed-by: Paul Szczepanek <paul.szczepanek@arm.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>